fix(ci): python-coverage-shards — pr_monitor.py exceeded 1500-line maintainability limit
Recent CI-repair-routing commits pushed src/awf/runtime/pr_monitor.py to 1510
lines, tripping test_first_party_code_files_stay_under_line_limit (max 1500).
Extract the wire-shape input dataclasses/enums into a new one-directional
pr_monitor_models module and re-export them (explicit __all__) so the pure
decision core drops to ~1313 lines while every historical
`from awf.runtime.pr_monitor import PRStatus` call site keeps resolving.
